refactor: more redux architecture
