+++ src/reducers/localChatReducer.js
@@ -5,7 +5,7 @@
 import Immutable from 'immutable'
 
 // Filter the data
-export function localChatStore (state = Immutable.List(), action) {
+export default function localChatReducer (state = Immutable.List(), action) {
   switch (action.type) {
     case 'ChatFromSimulator':
       return state.push(Immutable.Map(action.msg))
